[HOTFIX] Revert "[rollout] production rollout June 17th, 2020" (#1249)
This reverts commit 25d1a7bd43056b79974785c205307d3b20379694, reversing
changes made to 4bd9b613c4526a2facac4881144b4555bb04b3af.

+++ src/Maestro/Maestro.Web/Api/v2020_02_20/Controllers/ChannelsController.cs
@@ -76,7 +76,7 @@ namespace Maestro.Web.Api.v2020_02_20.Controllers
         }
 
         /// <summary>
-        ///   Gets a single <see cref="Channel"/>.
+        ///   Gets a single <see cref="Channel"/>, including all <see cref="ReleasePipeline"/> data.
         /// </summary>
         /// <param name="id">The id of the <see cref="Channel"/> to get</param>
         [HttpGet("{id}")]
@@ -85,6 +85,8 @@ namespace Maestro.Web.Api.v2020_02_20.Controllers
         public override async Task<IActionResult> GetChannel(int id)
         {
             Data.Models.Channel channel = await _context.Channels
+                .Include(ch => ch.ChannelReleasePipelines)
+                .ThenInclude(crp => crp.ReleasePipeline)
                 .Where(c => c.Id == id).FirstOrDefaultAsync();
 
             if (channel == null)
@@ -105,6 +107,7 @@ namespace Maestro.Web.Api.v2020_02_20.Controllers
         public override async Task<IActionResult> DeleteChannel(int id)
         {
             Data.Models.Channel channel = await _context.Channels
+                .Include(ch => ch.ChannelReleasePipelines)
                 .FirstOrDefaultAsync(c => c.Id == id);
 
             if (channel == null)
@@ -118,6 +121,13 @@ namespace Maestro.Web.Api.v2020_02_20.Controllers
                 return BadRequest(
                     new ApiError($"The channel with id '{id}' has associated subscriptions. " +
                     "Please remove these before removing this channel."));
+            }
+
+            if (channel.ChannelReleasePipelines != null && channel.ChannelReleasePipelines.Any())
+            {
+                return BadRequest(
+                    new ApiError($"The channel with id '{id}' has '{channel.ChannelReleasePipelines.Count()}' " +
+                    $"release pipeline(s) attached to it. Detach those release pipelines(s) first."));
             }
 
             _context.Channels.Remove(channel);
